fix compiler warnings

+++ Source/cmAuxSourceDirectoryCommand.cxx
@@ -49,7 +49,7 @@ bool cmAuxSourceDirectoryCommand::InitialPass(std::vector<std::string> const& ar
     size_t numfiles = dir.GetNumberOfFiles();
     for(size_t i =0; i < numfiles; ++i)
       {
-      std::string file = dir.GetFile(i);
+      std::string file = dir.GetFile(static_cast<unsigned long>(i));
       // Split the filename into base and extension
       std::string::size_type dotpos = file.rfind(".");
       if( dotpos != std::string::npos )
